move common to hatt
