+++ crates/ferrmion-core/src/encode/README.md
@@ -1,0 +1,5 @@
+This folder contains fermion-qubit encodings.
+
+- `encoding.rs`: Generic `MajoranaEncoding`.
+- `ternarytree.rs`: Methods to build general `TernaryTree`s as well as helper functions for Jordan-Wigner, Bravyi-Kitaev, Parity and JKMN encodings. 
+- `maxnto.rs`: `max_nto_symplectic_matrix` to define the `MaxNTO` encoding.s
