+++ crates/ferrmion-core/src/optimise/README.md
@@ -1,0 +1,6 @@
+This folder contains optimisation methods for fermion-qubit encodings.
+
+- `anneal.rs`: Simulated annealing
+- `hatt.rs`: Hamiltonian adaptive ternary tree
+- `topphatt.rs`: Hamiltonian adaptive ternary tree with top-p sampling
+  - This has some dependencies on `hatt.rs`.
