+++ crates/ferrmion-core/src/optimise/mod.rs
@@ -5,7 +5,6 @@
 
 mod anneal;
 pub use anneal::*;
-mod common;
 mod hatt;
 pub use hatt::*;
 mod topphatt;
